split mandatory.c into individual files
